+++ R_type.h
@@ -96,207 +96,10 @@ public:
     virtual ~R_type(){}
 };
 
-
-class Numeric_t:public R_type
-{
-    std::vector<double *> v;
-    void clear()
-    {
-        if (!isfake()){
-            for (unsigned int i = 0; i < v.size(); i++)
-            {
-                if (v[i] != NULL)
-                    delete v[i];
-            }
-        }
-        v.clear();
-        v.push_back(NULL);
-    }
-public:
-    ~Numeric_t() {clear();}
-    Numeric_t():R_type(Numeric){
-        v.push_back(NULL);
-        R_type::size = 0;
-    };
-    void add(double x)
-    {
-        double *tmp = new double (x);
-        v.push_back(tmp);
-        R_type::size ++;
-    }
-    void add() {
-        v.push_back(NULL);
-        R_type::size++;
-    }
-    void addaddr(double *x){
-        v.push_back(x);
-    }
-    Numeric_t(double x):R_type(Numeric)
-    {
-        v.push_back(NULL);
-        double *px = new double(x);
-        v.push_back(px);
-        R_type::size = 1;
-    }
-
-    void write()
-    {
-        for (unsigned int i = 1; i < v.size(); i++)
-        {
-            if (v[i] != NULL)
-                std::cout << *v[i] << ' ';
-            else
-                std::cout << "NULL ";
-        }
-    }
-
-    int Plus(R_type *);
-    int Minus(R_type *);
-    int Mul(R_type *);
-    int Div(R_type *);
-    int Eq(R_type *);
-    int Neq(R_type *);
-    int Bigger(R_type *);
-    int Beq(R_type *);
-    int Less(R_type *);
-    int Leq(R_type *);
-    int Unmin();
-    R_type *ToChar();
-    R_type *ToLog();
-    R_type *ToNum(){return this;}
-    int interval(R_type *);
-    int at(const std::set<unsigned int> &);
-    std::set<unsigned int> get_tuple();
-    double* kernel_at_num(int i) { return v[i];}
-    int in(R_type *);
-};
-
-
-
-class Character_t:public R_type{
-    std::vector<std::string *> v;
-    void clear()
-    {
-        if (!isfake()){
-            for (unsigned int i = 0; i < v.size(); i++)
-            {
-                if (v[i] != NULL)
-                    delete v[i];
-            }
-        }
-        v.clear();
-        v.push_back(NULL);
-        R_type::size = 0;
-    }
-public:
-    Character_t(std::string &s):R_type(Character){
-        v.push_back(NULL);
-        std::string *ps = new std::string(s);
-        v.push_back(ps);
-        R_type::size = 1;
-    }
-    Character_t():R_type(Character){v.push_back(NULL); R_type::size = 0;}
-    void add(){
-        v.push_back(NULL);
-        R_type::size++;
-    }
-    void add(std::string &s)
-    {
-        std::string *ps  = new std::string(s);
-        v.push_back(ps);
-        R_type::size++;
-    }
-    void addaddr(std::string *s)
-    {
-        v.push_back(s);
-        R_type::size++;
-    }
-    void write()
-    {
-        for (unsigned int i = 1; i < v.size(); i++)
-        {
-            if (v[i] != NULL)
-                std::cout << *v[i] << ' ';
-            else
-                std::cout << "NULL ";
-        }
-    }
-    int Beq(R_type *);
-    int Eq(R_type *);
-    int Neq(R_type *);
-    int Bigger(R_type *);
-    int Less(R_type *);
-    int Leq(R_type *);
-    int at(const std::set<unsigned int> &);
-    int in(R_type *);
-    std::string* kernel_at_char(int i){return v[i];}
-    R_type * ToNum(){throw "Cannot assign character to numeric";}
-    R_type * ToLog(){throw "Cannot assign character to logical";}
-    R_type * ToChar(){return this;}
-};
-
-
-class Logical_t:public R_type{
-    std::vector <bool *> v;
-    void clear()
-    {
-        if (!isfake()){
-            for (unsigned int i = 0; i < v.size(); i++)
-            {
-                if (v[i] != NULL)
-                    delete v[i];
-            }
-        }
-        v.clear();
-        v.push_back(NULL);
-        R_type::size = 0;
-    }
-public:
-    void add (bool a)
-    {
-        bool *tmp = new bool(a);
-        v.push_back(tmp);
-        R_type::size++;
-    }
-    void add(){ v.push_back(NULL); R_type::size++;}
-    void addaddr(bool *a){ v.push_back(a); R_type::size++;}
-    ~Logical_t()
-    {
-        clear();
-    }
-    Logical_t ():R_type(Logical){
-        R_type::size = 0;
-        v.push_back(NULL);
-    }
-    Logical_t (bool b) :R_type(Logical)
-    {
-        v.push_back(NULL);
-        bool *tmp = new bool(b);
-        v.push_back(tmp);
-        R_type::size = 1;
-    }
-    void write()
-    {
-        for (unsigned int i = 1; i < v.size(); i++)
-        {
-            if (v[i] != NULL)
-                (*v[i]) ?  std::cout << "TRUE" << ' ' : std::cout <<  "FALSE" << ' ';
-            else
-                std::cout << "NULL ";
-        }
-    }
-    int Not();
-    bool * kernel_at_log(int i){return v[i];}
-    int And(R_type * );
-    int Or(R_type * );
-    int in(R_type *);
-    std::set<unsigned int> get_tuple();
-    int at(const std::set<unsigned int> &);
-    R_type * ToChar();
-    R_type * ToNum();
-    R_type * ToLog(){return this;}
-};
-
+//forward declaration of types
+class Numeric_t;
+class Logical_t;
+class Character_t;
 
 R_type* change_type(R_type *a, kind_of_type t)
 {
@@ -311,3 +114,5 @@ R_type* change_type(R_type *a, kind_of_type t)
     }
     return ans;
 }
+
+#endif
